Clean up legacy code
